refactor(history,shop): move db to a separate package
