+++ .github/workflows/build.yml
@@ -42,3 +42,12 @@ jobs:
 
       - name: Build logger
         uses: ./packages/logger
+
+  build-db:
+    runs-on: ubuntu-latest
+    steps:
+      - name: Checkout code
+        uses: actions/checkout@v3
+
+      - name: Build db
+        uses: ./packages/db
