+++ packages/db/.prettierignore
@@ -1,0 +1,5 @@
+node_modules
+dist
+yarn.lock
+package-lock.json
+public
