+++ packages/db/.prettierrc
@@ -1,0 +1,6 @@
+{
+  "trailingComma": "all",
+  "semi": true,
+  "singleQuote": true,
+  "quoteProps": "consistent"
+}
