+++ packages/db/action.yml
@@ -1,0 +1,25 @@
+name: Build db
+
+description: |-
+  Steps to build db package
+
+runs:
+  using: 'composite'
+  steps:
+    - name: Setup environment
+      uses: ./.github/actions/setup
+
+    - name: Run lint check
+      run: yarn lint
+      shell: bash
+      working-directory: ${{github.action_path}}
+
+    - name: Run test
+      run: yarn test
+      shell: bash
+      working-directory: ${{github.action_path}}
+
+    - name: Run type-check
+      run: yarn type-check
+      shell: bash
+      working-directory: ${{github.action_path}}
