+++ packages/db/index.ts
@@ -1,0 +1,1 @@
+export { query } from './src/query';
