+++ services/shop/package.json
@@ -19,14 +19,11 @@
     "validate": "yarn lint && yarn type-check && yarn test"
   },
   "dependencies": {
-    "@koa/cors": "3.3.0",
     "@koa/router": "^10.1.1",
-    "@nab/logger": "^0.1.0",
+    "@nab/db": "^0.1.0",
     "@nab/http": "^0.1.0",
+    "@nab/logger": "^0.1.0",
     "axios": "^0.26.1",
-    "koa": "^2.13.4",
-    "koa-bodyparser": "^4.3.0",
-    "koa-helmet": "^6.1.0",
     "lodash": "^4.17.21",
     "pg": "^8.7.3",
     "ts-node": "^10.7.0",
@@ -34,10 +31,6 @@
   },
   "devDependencies": {
     "@types/jest": "^27.4.1",
-    "@types/koa": "^2.13.4",
-    "@types/koa-bodyparser": "^4.3.7",
-    "@types/koa-helmet": "^6.0.4",
-    "@types/koa__cors": "^3.1.1",
     "@types/koa__router": "^8.0.11",
     "@types/lodash": "^4.14.181",
     "@types/node": "^17.0.23",
